// apps/backend/src/modules/human-resources/types/save-motorizer.interface.ts
export interface SaveMotorizer {
  original_doc?: string
  id?: number
  name: string
  email: string
  doc_type: null | string
  doc_number: null | string
  vehicle: null | string
  plate: null | string
  phone: string
  shift_hired: string
  status: number
  store_name?: string
  store_id: string
  store_code: string
}